+++ src/test/java/blackjack/domain/BettingTableTest.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -8,9 +8,6 @@ import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import blackjack.domain.Betting;
-import blackjack.domain.BettingTable;
-import blackjack.domain.PlayRecord;
 import blackjack.domain.participant.Name;
 
 public class BettingTableTest {
